Fix CPF to mandatory and consume AuthService

+++ PeopleService/src/main/java/org/uniara/PeopleService/controller/PersonController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import org.uniara.PeopleService.DTO.CreatePersonDTO;
 import org.uniara.PeopleService.DTO.UpdatePersonDTO;
 import org.uniara.PeopleService.constant.Constant;
+import org.uniara.PeopleService.consumers.AuthConsumer;
 import org.uniara.PeopleService.model.Person;
 import org.uniara.PeopleService.service.PersonService;
 
@@ -19,13 +20,25 @@ public class PersonController {
     @Autowired
     private PersonService personService;
 
+    private AuthConsumer authConsumer = new AuthConsumer();
+
     @GetMapping(Constant.API_PEOPLE)
-    public ResponseEntity<List<Person>> findAll(/*@RequestHeader("Authorization") String token*/) {
+    public ResponseEntity<List<Person>> findAll(@RequestHeader("Authorization") String token) {
+
+        if (!authConsumer.isAuthenticated(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(personService.findAll());
     }
 
     @PostMapping(Constant.API_PEOPLE)
-    public ResponseEntity<Person> save(/*@RequestHeader("Authorization") String token, */@RequestBody CreatePersonDTO dto) {
+    public ResponseEntity<Person> save(@RequestHeader("Authorization") String token, @RequestBody CreatePersonDTO dto) {
+
+        if (!authConsumer.isAuthenticated(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         Person person = new Person();
 
         if (person.getCreatedAt() == null) {
@@ -33,32 +46,46 @@ public class PersonController {
         }
 
         person.setName(dto.getName());
+        person.setCpf(dto.getCpf());
         person.setPhoneNumber(dto.getPhoneNumber());
 
         return ResponseEntity.status(HttpStatus.CREATED).body(personService.save(person));
     }
 
     @PutMapping(Constant.API_PEOPLE)
-    public ResponseEntity<Person> update(/*@RequestHeader("Authorization") String token, */@RequestBody UpdatePersonDTO dto) {
-        Optional<Person> customer = personService.findById(dto.getId());
+    public ResponseEntity<Person> update(@RequestHeader("Authorization") String token, @RequestBody UpdatePersonDTO dto) {
+        if (!authConsumer.isAuthenticated(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
-        if (customer.isEmpty()) {
+        Optional<Person> person = personService.findById(dto.getId());
+
+        if (person.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
 
-        customer.get().setName(dto.getName());
-        customer.get().setPhoneNumber(dto.getPhoneNumber());
+        person.get().setName(dto.getName());
+        person.get().setCpf(dto.getCpf());
+        person.get().setPhoneNumber(dto.getPhoneNumber());
 
-        return ResponseEntity.ok(personService.save(customer.get()));
+        return ResponseEntity.ok(personService.save(person.get()));
     }
 
     @GetMapping(Constant.API_PEOPLE + "/{id}")
-    public ResponseEntity<Optional<Person>> findById(/*@RequestHeader("Authorization") String token, */@PathVariable("id") String id) {
+    public ResponseEntity<Optional<Person>> findById(@RequestHeader("Authorization") String token, @PathVariable("id") String id) {
+        if (!authConsumer.isAuthenticated(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         return ResponseEntity.ok(personService.findById(id));
     }
 
     @DeleteMapping(Constant.API_PEOPLE + "/{id}")
-    public ResponseEntity<Person> deleteById(/*@RequestHeader("Authorization") String token, */@PathVariable("id") String id) {
+    public ResponseEntity<Person> deleteById(@RequestHeader("Authorization") String token, @PathVariable("id") String id) {
+        if (!authConsumer.isAuthenticated(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         personService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
